Fixed issue with frontier.py (ignore revert)
In order to ensure consistent hashing, frontier.py always strips the right slash for a URL before calculating its hash. The problem is that this right slash is crucial information to the URL- it tells the browser/client whether we can go any deeper than we already are. We shouldn't be the ones adding that slash, because there's the case that we aren't meant to be going deeper, which could cause an infinite trap (and did cause one last night).